requirements.txt
alembic==1.13.1
async-timeout==4.0.3
Authlib==1.3.0
blinker==1.7.0
cachelib==0.12.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
DateTime==5.5
dnspython==2.6.1
email_validator==2.1.1
Flask==3.0.2
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-OAuthlib==0.9.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
oauthlib==2.1.0
psycopg==3.1.18
psycopg-binary==3.1.18
psycopg2-binary==2.9.9
pycparser==2.21
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.3
requests==2.31.0
requests-oauthlib==1.1.0
six==1.16.0
spotipy==2.23.0
SQLAlchemy==2.0.28
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==3.0.2
WTForms==3.1.2
zope.interface==6.3
